+++ packages/astro/test/test-utils.js
@@ -41,7 +41,8 @@ process.env.ASTRO_TELEMETRY_DISABLED = true;
  * @property {(inlineConfig?: Parameters<typeof dev>[0]) => ReturnType<typeof dev>} startDevServer
  * @property {typeof preview} preview
  * @property {() => Promise<void>} clean
- * @property {() => Promise<App>} loadTestAdapterApp
+ * @property {(streaming: boolean) => Promise<App>} loadTestAdapterApp
+ * @property {(streaming: boolean) => Promise<App>} loadSelfAdapterApp
  * @property {() => Promise<(req: NodeRequest, res: NodeResponse) => void>} loadNodeAdapterHandler
  * @property {(timeout?: number) => Promise<void>} onNextDataStoreChange
  * @property {typeof check} check
@@ -271,15 +272,15 @@ export async function loadFixture(inlineConfig) {
 			const { handler } = await import(url);
 			return handler;
 		},
-		loadTestAdapterApp: async (streaming, entryName = 'entry.mjs') => {
-			const url = new URL(`./server/${entryName}?id=${fixtureId}`, config.outDir);
+		loadTestAdapterApp: async (streaming) => {
+			const url = new URL(`./server/${config.build.serverEntry}?id=${fixtureId}`, config.outDir);
 			const { createApp, manifest } = await import(url);
 			const app = createApp(streaming);
 			app.manifest = manifest;
 			return app;
 		},
-		loadSelfAdapterApp: async (streaming, entryName = 'entry.mjs') => {
-			const url = new URL(`./server/${entryName}?id=${fixtureId}`, config.outDir);
+		loadSelfAdapterApp: async (streaming) => {
+			const url = new URL(`./server/${config.build.serverEntry}?id=${fixtureId}`, config.outDir);
 			const { createApp } = await import(url);
 			return createApp(streaming);
 		},